Refactor auth performance monitor into modular components
