+++ V2_COMPLIANCE_PROGRESS_TRACKER.md
@@ -174,11 +174,11 @@
 > **Note:** Remaining contract descriptions include current line counts for context only. There is no strict LOC target—deliver clean, production-ready, tested modules that honor SRP and SOLID principles.
 
 ### 🎉 **CRITICAL PRIORITY - COMPLETED! (0 files over 800 lines)**
-**All critical violations have been successfully resolved!** 
+**All critical violations have been successfully resolved!**
 
 **Previously resolved files:**
 1. **`src/services/financial/portfolio/tracking.py`** - ✅ **REFACTORED** (937 → 32 lines)
-2. **`src/core/health/alerting/manager.py`** - ✅ **REFACTORED** (910 → 233 lines)  
+2. **`src/core/health/alerting/manager.py`** - ✅ **REFACTORED** (910 → 233 lines)
 3. **`src/services/financial/unified_financial_api.py`** - ✅ **REFACTORED** (872 → 730 lines)
 4. **`src/services/integrated_agent_coordinator.py`** - ✅ **REFACTORED** (846 → 84 lines)
 5. **`src/core/health/metrics/collector.py`** - ✅ **CONSOLIDATED** (839 → consolidated)
@@ -486,6 +486,8 @@ Code quality, organization, and maintainability over strict line count limits
 16. ✅ **CRIT-017**: OSRS AI Agent Test Suite (900 → 160 lines)
     - Split monolithic tests into game setup, agent, scenario, and validation modules
     - Added orchestrator for targeted execution
+
+17. ✅ **MAJOR-010**: Auth Performance Monitor modularized into core, metrics, reporting, and config modules
 
 ### Impact:
 - **Total lines reduced**: 14,962 → 3,466 lines (76.8% reduction)
